requirements.txt
annotated-types==0.7.0
asn1crypto==1.5.1
boto3==1.38.31
botocore==1.38.31
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.3
cryptography==45.0.3
cycler==0.12.1
filelock==3.18.0
fonttools==4.59.0
idna==3.10
jmespath==1.0.1
kiwisolver==1.4.9
matplotlib==3.10.5
numpy==2.2.6
packaging==25.0
pandas==2.3.0
pillow==11.3.0
platformdirs==4.3.8
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.13.0
seaborn==0.13.2
six==1.17.0
snowflake-connector-python==3.15.0
snowflake.core==1.5.1
sortedcontainers==2.4.0
tomlkit==0.13.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
